+++ src/main/java/net/spell_engine/api/spell/Spell.java
@@ -36,6 +36,8 @@ public class Spell {
         public int channel_ticks = 0;
         public String animation;
         public boolean animates_ranged_weapon = false;
+        /// Default `0.2` matches the same as movement speed during vanilla item usage (such as bow)"
+        public float movement_speed = 0.2F;
         public Sound start_sound;
         public Sound sound;
         public ParticleBatch[] particles = new ParticleBatch[]{};