+++ src/main/java/net/spell_engine/config/ServerConfig.java
@@ -9,8 +9,8 @@ import java.util.LinkedHashMap;
 
 @Config(name = "server")
 public class ServerConfig implements ConfigData { public ServerConfig() {}
-    @Comment("Default `0.2` matches the same as movement speed during vanilla item usage (such as bow)")
-    public float movement_speed_while_casting_spell = 0.2F;
+    @Comment("Applied as multiplier on top of spell.cast.movement_speed. Default value of 1.0 means no change.")
+    public float movement_multiplier_speed_while_casting = 1F;
     @Comment("Allow spells to bypass invulnerability frames. This is required in order for high attack frequency spells (such as beams) to work.")
     public boolean bypass_iframes = true;
     @Comment("Spell haste reduces the cooldown time of abilities")